python/requirements.txt
websockets
sounddevice
numpy
openai-whisper 